requirements.txt
# -----------------------
# Core web stack
# -----------------------
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
Flask-WTF==1.2.1
WTForms==3.1.2
WTForms-SQLAlchemy==0.4.2
Jinja2==3.1.3
Werkzeug==3.0.2
itsdangerous==2.1.2
gunicorn==22.0.0

# -----------------------
# Database / search
# -----------------------
SQLAlchemy==2.0.29
# Use a prebuilt wheel in Linux containers to avoid pg_config issues
psycopg2-binary==2.9.10; platform_system != "Windows"
# Use the normal package on Windows (where pg_config can come from the installer)
psycopg2==2.9.10; platform_system == "Windows"
pgvector==0.4.1

# -----------------------
# Data / utils
# -----------------------
numpy==1.26.4
pandas==2.2.1
python-dotenv==1.1.0
requests==2.31.0
redis==5.0.1

# -----------------------
# Documents & PDF
# -----------------------
python-docx==1.1.0
python-pptx==0.6.23        # provides 'pptx' module
PyMuPDF==1.26.3            # provides 'fitz' module
pdfplumber==0.11.0
pdf2image==1.17.0
pypdfium2==4.28.0
Pillow==10.3.0
openpyxl==3.1.2            # keep 3.1.2 to avoid resolver conflicts you hit with 3.1.5

# -----------------------
# Windows-only (guarded)
# These enable COM/Word conversion when you run natively on Windows.
# They are skipped in Docker/Linux thanks to the markers below.
# -----------------------
docx2pdf==0.1.8; platform_system == "Windows"
pywin32==306; platform_system == "Windows"
comtypes==1.4.0; platform_system == "Windows"

# -----------------------
sentence-transformers==2.6.1
transformers==4.39.3
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
tqdm==4.66.2
spacy==3.7.4
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1   # makes fuzzywuzzy fast, removes the warning
rapidfuzz==3.9.7
thefuzz==0.22.1
openai==1.43.0
psutil==6.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
datasets==2.19.1
#py3exiv2
pyodbc

